+++ packages/ui/src/components/Badge.tsx
@@ -1,6 +1,5 @@
 import { type BadgeProps as BadgePropsUi, Badge as BadgeUi } from 'theme-ui';
-
-import { Icon, IconType } from './Icon';
+import { Icon, IconType } from '@ttoss/react-icons';
 
 export type BadgeProps = BadgePropsUi & {
   icon?: IconType;