+++ terezinha-farm/app/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.36](https://github.com/ttoss/ttoss/compare/@terezinha-farm/app@0.5.35...@terezinha-farm/app@0.5.36) (2023-08-15)
+
+**Note:** Version bump only for package @terezinha-farm/app
+
+## [0.5.35](https://github.com/ttoss/ttoss/compare/@terezinha-farm/app@0.5.34...@terezinha-farm/app@0.5.35) (2023-08-15)
+
+**Note:** Version bump only for package @terezinha-farm/app
+
 ## [0.5.34](https://github.com/ttoss/ttoss/compare/@terezinha-farm/app@0.5.33...@terezinha-farm/app@0.5.34) (2023-08-11)
 
 **Note:** Version bump only for package @terezinha-farm/app